requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
camelot-py==0.7.3
chardet==3.0.4
Click==7.0
cycler==0.10.0
et-xmlfile==1.0.1
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10
kiwisolver==1.1.0
MarkupSafe==1.1.0
matplotlib==3.1.1
numpy==1.17.0
opencv-python==4.1.0.25
openpyxl==2.6.2
pandas==0.25.0
pdfminer.six==20181108
Pillow==5.4.1
pycryptodome==3.8.2
pyparsing==2.4.2
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.8
Werkzeug==0.14.1
wget==3.2
